Separate channel controller to it's own file
