+++ kafka/channel/pkg/reconciler/controller/controller.go
@@ -1,0 +1,131 @@
+/*
+Copyright 2020 The Knative Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package controller
+
+import (
+	"context"
+
+	"github.com/kelseyhightower/envconfig"
+	"go.uber.org/zap"
+	"k8s.io/client-go/tools/cache"
+
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	kubeclient "knative.dev/pkg/client/injection/kube/client"
+	"knative.dev/pkg/client/injection/kube/informers/apps/v1/deployment"
+	"knative.dev/pkg/client/injection/kube/informers/core/v1/endpoints"
+	"knative.dev/pkg/client/injection/kube/informers/core/v1/service"
+	"knative.dev/pkg/client/injection/kube/informers/core/v1/serviceaccount"
+	"knative.dev/pkg/client/injection/kube/informers/rbac/v1/rolebinding"
+
+	"knative.dev/eventing/pkg/logging"
+	"knative.dev/eventing/pkg/reconciler"
+	"knative.dev/pkg/configmap"
+	"knative.dev/pkg/controller"
+	"knative.dev/pkg/system"
+
+	kafkaclientsetinjection "knative.dev/eventing-contrib/kafka/channel/pkg/client/injection/client"
+	"knative.dev/eventing-contrib/kafka/channel/pkg/client/injection/informers/messaging/v1alpha1/kafkachannel"
+)
+
+// NewController initializes the controller and is called by the generated code.
+// Registers event handlers to enqueue events.
+func NewController(
+	ctx context.Context,
+	cmw configmap.Watcher,
+) *controller.Impl {
+	logger := logging.FromContext(ctx)
+
+	kafkaChannelInformer := kafkachannel.Get(ctx)
+	deploymentInformer := deployment.Get(ctx)
+	endpointsInformer := endpoints.Get(ctx)
+	serviceAccountInformer := serviceaccount.Get(ctx)
+	roleBindingInformer := rolebinding.Get(ctx)
+	serviceInformer := service.Get(ctx)
+
+	kafkaChannelClientSet := kafkaclientsetinjection.Get(ctx)
+
+	r := &Reconciler{
+		Base:            reconciler.NewBase(ctx, controllerAgentName, cmw),
+		systemNamespace: system.Namespace(),
+
+		kafkachannelLister:   kafkaChannelInformer.Lister(),
+		kafkachannelInformer: kafkaChannelInformer.Informer(),
+		deploymentLister:     deploymentInformer.Lister(),
+		serviceLister:        serviceInformer.Lister(),
+		endpointsLister:      endpointsInformer.Lister(),
+		serviceAccountLister: serviceAccountInformer.Lister(),
+		roleBindingLister:    roleBindingInformer.Lister(),
+		kafkaClientSet:       kafkaChannelClientSet,
+	}
+
+	env := &envConfig{}
+	if err := envconfig.Process("", env); err != nil {
+		r.Logger.Panicf("unable to process Kafka channel's required environment variables: %v", err)
+	}
+
+	if env.Image == "" {
+		r.Logger.Panic("unable to process Kafka channel's required environment variables (missing DISPATCHER_IMAGE)")
+	}
+
+	r.dispatcherImage = env.Image
+
+	r.impl = controller.NewImpl(r, r.Logger, ReconcilerName)
+
+	// Get and Watch the Kakfa config map and dynamically update Kafka configuration.
+	if _, err := kubeclient.Get(ctx).CoreV1().ConfigMaps(system.Namespace()).Get("config-kafka", metav1.GetOptions{}); err == nil {
+		cmw.Watch("config-kafka", r.updateKafkaConfig)
+	} else if !apierrors.IsNotFound(err) {
+		logger.With(zap.Error(err)).Fatal("Error reading ConfigMap 'config-kafka'")
+	}
+
+	r.Logger.Info("Setting up event handlers")
+	kafkaChannelInformer.Informer().AddEventHandler(controller.HandleAll(r.impl.Enqueue))
+
+	// Set up watches for dispatcher resources we care about, since any changes to these
+	// resources will affect our Channels. So, set up a watch here, that will cause
+	// a global Resync for all the channels to take stock of their health when these change.
+	filterFn := controller.FilterWithName(dispatcherName)
+
+	// Call GlobalResync on kafkachannels.
+	grCh := func(obj interface{}) {
+		r.impl.GlobalResync(r.kafkachannelInformer)
+	}
+
+	deploymentInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
+		FilterFunc: filterFn,
+		Handler:    controller.HandleAll(grCh),
+	})
+	serviceInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
+		FilterFunc: filterFn,
+		Handler:    controller.HandleAll(grCh),
+	})
+	endpointsInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
+		FilterFunc: filterFn,
+		Handler:    controller.HandleAll(grCh),
+	})
+	serviceAccountInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
+		FilterFunc: filterFn,
+		Handler:    controller.HandleAll(grCh),
+	})
+	roleBindingInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
+		FilterFunc: filterFn,
+		Handler:    controller.HandleAll(grCh),
+	})
+
+	return r.impl
+}
